more error handling, better dry-run, implied --regex on --new, --new without --find, {{index}} token, more tests

+++ cli.js
@@ -13,29 +13,39 @@ function green(txt){
     return "\x1b[32m" + txt + "\x1b[0m";
 }
 
-var optionSet = new Thing()
-    .mixIn(new rename.RenameOptions(), "rename")
-    .define({ 
-        name: "files",
-        type: Array,
-        required: true,
-        defaultOption: true,
-        groups: ["rename"],
-        valueTest: [
-            function(files){ return files.every(fs.existsSync); },
-            function(files){ return files.length > 0; }
-        ],
-        valueFailMsg: "Must be at least one file, and all must exist"
-    })
-    .define({ name: "dry-run", type: "boolean", alias: "d" })
-    .on("error", function(err){
-        console.error(red("Error: ") + err.message);
-        process.exit(1);
-    })
-    .set(process.argv);
+/*
+TODO: Tidy error handling
+*/
+var optionSet;
+try {
+    optionSet = new Thing()
+        .mixIn(new rename.RenameOptions(), "rename")
+        .define({ 
+            name: "files",
+            type: Array,
+            required: true,
+            defaultOption: true,
+            groups: ["rename"],
+            valueTest: [
+                function(files){ return files.every(fs.existsSync); },
+                function(files){ return files.length > 0; }
+            ],
+            valueFailMsg: "Must be at least one file, and all must exist"
+        })
+        .define({ name: "dry-run", type: "boolean", alias: "d" })
+        .on("error", function(err){
+            console.error(red("Error: ") + err.message);
+            process.exit(1);
+        })
+        .set(process.argv);
+} catch (e){
+    console.error(red("Invalid argument: " + e.message));
+    process.exit(1);
+}
 
 if (optionSet.valid){
-    var results;
+    var results,
+        newFilenames = [];
     try {
         results = rename.rename(optionSet.where({ group: "rename" }));
     } catch (e){
@@ -46,14 +56,12 @@ if (optionSet.valid){
         if (result.before === result.after || !result.after ){
             console.log(red("no change: ") + result.before);
         } else {
-            console.log(green("rename: ") + result.before, "->", result.after);
-            if (optionSet["dry-run"]){
-                // do nothing else
-            } else if (!fs.existsSync(result.after)){
-                fs.renameSync(result.before, result.after);
+            if (fs.existsSync(result.after) || newFilenames.indexOf(result.after) > -1){
+                console.log(red("no change: "), result.before, "->", result.after, red("(file exists)"));
             } else {
-                console.error("a file by that new name already exists: " + result.after);
-                return;
+                if (!optionSet["dry-run"]) fs.renameSync(result.before, result.after);
+                newFilenames.push(result.after);
+                console.log(green("rename: ") + result.before, "->", result.after);
             }
         }
     });
